+++ layout/src/main/java/com/vlkan/log4j2/logstash/layout/LogstashLayout.java
@@ -25,6 +25,7 @@ import org.apache.logging.log4j.core.util.datetime.FastDateFormat;
 import org.apache.logging.log4j.util.Supplier;
 
 import java.io.IOException;
+import java.lang.reflect.Method;
 import java.nio.ByteBuffer;
 import java.nio.charset.Charset;
 import java.nio.charset.StandardCharsets;
@@ -44,6 +45,8 @@ public class LogstashLayout implements Layout<String> {
 
     private static final byte[] EMPTY_OBJECT_JSON_BYTES = "{}".getBytes(CHARSET);
 
+    private static final ObjectMapper DEFAULT_OBJECT_MAPPER = new ObjectMapper();
+
     private final TemplateResolver<LogEvent> eventResolver;
 
     private final byte[] lineSeparatorBytes;
@@ -53,7 +56,7 @@ public class LogstashLayout implements Layout<String> {
     private LogstashLayout(Builder builder) {
 
         // Create StackTraceElement resolver.
-        ObjectMapper objectMapper = new ObjectMapper();
+        ObjectMapper objectMapper = createObjectMapper(builder.objectMapper, builder.objectMapperFactoryMethod);
         StrSubstitutor substitutor = builder.config.getStrSubstitutor();
         TemplateResolver<StackTraceElement> stackTraceElementObjectResolver = null;
         if (builder.stackTraceEnabled) {
@@ -94,6 +97,22 @@ public class LogstashLayout implements Layout<String> {
                 builder.emptyPropertyExclusionEnabled,
                 builder.maxStringLength);
 
+    }
+
+    private static ObjectMapper createObjectMapper(ObjectMapper objectMapper, String objectMapperFactoryMethod) {
+        if (objectMapperFactoryMethod != null) {
+            try {
+                int splitterIndex = objectMapperFactoryMethod.lastIndexOf('.');
+                String className = objectMapperFactoryMethod.substring(0, splitterIndex);
+                String methodName = objectMapperFactoryMethod.substring(splitterIndex + 1);
+                Class<?> clazz = Class.forName(className);
+                Method method = clazz.getMethod(methodName);
+                return (ObjectMapper) method.invoke(null);
+            } catch (Exception error) {
+                throw new RuntimeException(error);
+            }
+        }
+        return objectMapper;
     }
 
     private static String readEventTemplate(Builder builder) {
@@ -244,6 +263,12 @@ public class LogstashLayout implements Layout<String> {
         @PluginBuilderAttribute
         private int maxStringLength = 0;
 
+        @PluginBuilderAttribute
+        private ObjectMapper objectMapper = DEFAULT_OBJECT_MAPPER;
+
+        @PluginBuilderAttribute
+        private String objectMapperFactoryMethod;
+
         private Builder() {
             // Do nothing.
         }
@@ -392,6 +417,24 @@ public class LogstashLayout implements Layout<String> {
             return this;
         }
 
+        public ObjectMapper getObjectMapper() {
+            return objectMapper;
+        }
+
+        public Builder setObjectMapper(ObjectMapper objectMapper) {
+            this.objectMapper = objectMapper;
+            return this;
+        }
+
+        public String getObjectMapperFactoryMethod() {
+            return objectMapperFactoryMethod;
+        }
+
+        public Builder setObjectMapperFactoryMethod(String objectMapperFactoryMethod) {
+            this.objectMapperFactoryMethod = objectMapperFactoryMethod;
+            return this;
+        }
+
         @Override
         public LogstashLayout build() {
             validate();
@@ -412,6 +455,9 @@ public class LogstashLayout implements Layout<String> {
             }
             Validate.isTrue(maxByteCount > 0, "maxByteCount requires a non-zero positive integer");
             Validate.isTrue(maxStringLength >= 0, "maxStringLength requires a positive integer");
+            Validate.isTrue(
+                    objectMapper != null || objectMapperFactoryMethod != null,
+                    "both objectMapper and objectMapperFactoryMethod are null");
         }
 
         @Override